Increase key size from 64K to 4G

+++ arena.go
@@ -62,7 +62,7 @@ func (a *Arena) Reset() {
 	atomic.StoreUint32(&a.n, 1)
 }
 
-func (a *Arena) Alloc(size uint16, align Align) (uint32, error) {
+func (a *Arena) Alloc(size uint32, align Align) (uint32, error) {
 	// Pad the allocation with enough bytes to ensure the requested alignment.
 	padded := uint32(size) + uint32(align)
 
@@ -76,12 +76,12 @@ func (a *Arena) Alloc(size uint16, align Align) (uint32, error) {
 	return offset, nil
 }
 
-func (a *Arena) GetBytes(offset uint32, size uint16) []byte {
+func (a *Arena) GetBytes(offset uint32, size uint32) []byte {
 	if offset == 0 {
 		return nil
 	}
 
-	return a.buf[offset : offset+uint32(size)]
+	return a.buf[offset : offset+size]
 }
 
 func (a *Arena) GetPointer(offset uint32) unsafe.Pointer {
